Decide whether to test `INSERT IGNORE` and `DELETE IGNORE` based on dialects and fix the failing Oracle "test CRUD with Statements" tests

+++ integrated/src/main/kotlin/com/huanshankeji/exposedvertxsqlclient/Examples.kt
@@ -56,8 +56,8 @@ suspend fun examples(vertx: Vertx) {
     // put in `Vertx.executeBlocking` or `Dispatchers.IO` if needed
     createTablesWithExposedTransaction()
 
-    crudWithStatements(databaseClient)
-    crudExtensions(databaseClient)
+    crudWithStatements(databaseClient, true, false)
+    crudExtensions(databaseClient, true, false)
 }
 
 // for copying to README.md
@@ -97,15 +97,19 @@ suspend fun withTables(block: suspend () -> Unit) {
     }
 }
 
-suspend fun crudWithStatements(databaseClient: DatabaseClient<*>) {
+suspend fun crudWithStatements(
+    databaseClient: DatabaseClient<*>, dialectSupportsInsertIgnore: Boolean, dialectSupportsDeleteIgnore: Boolean
+) {
     val insertRowCount = databaseClient.executeUpdate(buildStatement { Examples.insert { it[name] = "A" } })
     assert(insertRowCount == 1)
     // `executeSingleUpdate` function requires that there is only 1 row updated and returns `Unit`.
     databaseClient.executeSingleUpdate(buildStatement { Examples.insert { it[name] = "B" } })
     // `executeSingleOrNoUpdate` requires that there is 0 or 1 row updated and returns `Boolean`.
-    val isInserted =
-        databaseClient.executeSingleOrNoUpdate(buildStatement { Examples.insertIgnore { it[name] = "B" } })
-    assert(isInserted)
+    if (dialectSupportsInsertIgnore) {
+        val isInserted =
+            databaseClient.executeSingleOrNoUpdate(buildStatement { Examples.insertIgnore { it[name] = "B" } })
+        assert(isInserted)
+    }
 
     val updateRowCount =
         databaseClient.executeUpdate(buildStatement { Examples.update({ Examples.id eq 1 }) { it[name] = "AA" } })
@@ -117,15 +121,22 @@ suspend fun crudWithStatements(databaseClient: DatabaseClient<*>) {
     assert(exampleName == "AA")
 
     databaseClient.executeSingleUpdate(buildStatement { Examples.deleteWhere { id eq 1 } })
-    // not supported by PostgreSQL
-    //databaseClient.executeSingleUpdate(buildStatement { Examples.deleteIgnoreWhere { id eq 2 } })
+    if (dialectSupportsDeleteIgnore) {
+        val isDeleted =
+            databaseClient.executeSingleOrNoUpdate(buildStatement { Examples.deleteIgnoreWhere { id eq 2 } })
+        assert(isDeleted)
+    }
 }
 
 @OptIn(ExperimentalEvscApi::class)
-suspend fun crudExtensions(databaseClient: DatabaseClient<*>) {
+suspend fun crudExtensions(
+    databaseClient: DatabaseClient<*>, dialectSupportsInsertIgnore: Boolean, dialectSupportsDeleteIgnore: Boolean
+) {
     databaseClient.insert(Examples) { it[name] = "A" }
-    val isInserted = databaseClient.insertIgnore(Examples) { it[name] = "B" }
-    assert(isInserted)
+    if (dialectSupportsInsertIgnore) {
+        val isInserted = databaseClient.insertIgnore(Examples) { it[name] = "B" }
+        assert(isInserted)
+    }
 
     val updateRowCount = databaseClient.update(Examples, { Examples.id eq 1 }) { it[name] = "AA" }
     assert(updateRowCount == 1)
@@ -142,9 +153,9 @@ suspend fun crudExtensions(databaseClient: DatabaseClient<*>) {
 
     val deleteRowCount1 = databaseClient.deleteWhere(Examples) { id eq 1 }
     assert(deleteRowCount1 == 1)
-    // not supported by PostgreSQL
-    /*
-    val deleteRowCount2 = databaseClient.deleteIgnoreWhere(Examples) { id eq 2 }
-    assert(deleteRowCount2 == 1)
-    */
+
+    if (dialectSupportsDeleteIgnore) {
+        val deleteRowCount2 = databaseClient.deleteIgnoreWhere(Examples) { id eq 2 }
+        assert(deleteRowCount2 == 1)
+    }
 }
